+++ tests/ext/test_sql.py
@@ -1,8 +1,10 @@
 import pytest
 from webtest import TestApp
 from pyramid.config import Configurator
-from resource_sql import BalloonModel
+from pyramid.testing import DummyRequest
 from pyramid_sqlalchemy import Session
+from resource_sql import BalloonModel
+from resource_sql import BalloonResource
 
 
 def make_app(config):
@@ -29,3 +31,14 @@ def test_known_id():
     app = make_app(config)
     r = app.get('/balloons/%s' % balloon.id)
     assert r.json['figure'] == u'Giraffe'
+
+
+@pytest.mark.usefixtures('sql_session')
+def test_update_instance():
+    balloon = BalloonModel(figure=u'Giraffe')
+    Session.add(balloon)
+    Session.flush()
+    request = DummyRequest(matchdict={'id': balloon.id})
+    resource = BalloonResource(request)
+    resource.update_from_dict({'figure': u'Elephant'})
+    assert balloon.figure == u'Elephant'